+++ tiled/authn_database/migrations/versions/27e069ba3bf5_add_deletion_scopes_to_default_roles.py
@@ -18,7 +18,7 @@ depends_on = None
 
 
 ROLES = ["admin", "user"]
-NEW_SCOPES = ["delete:metadata", "delete:data"]
+NEW_SCOPES = ["delete:revision", "delete:node"]
 
 
 def upgrade():